requirements.txt
# DS requirements
cffi==1.11.5
gevent==1.4.0
greenlet==0.4.15
hupper==1.4.2
LEPL==5.1.3
PasteDeploy==2.0.1
plaster==1.0
plaster-pastedeploy==0.6
pycparser==2.19
PyNaCl==1.3.0
pyramid==1.10.2
pyramid-exclog==1.0
pytz==2018.9
repoze.lru==0.7
six==1.12.0
translationstring==1.3
venusian==1.2.0
WebOb==1.8.5
zope.deprecation==4.4.0
zope.interface==4.6.0

git+https://github.com/ProzorroUKR/rfc6266.git
git+https://github.com/ProzorroUKR/chaussette.git@gevent_python3

# Storage requirements
debtcollector==1.20.0
iso8601==0.1.12
keystoneauth1==3.11.2
msgpack==0.6.1
netaddr==0.7.19
netifaces==0.10.9
os-service-types==1.5.0
oslo.config==6.8.0
oslo.i18n==3.23.0
oslo.serialization==2.28.1
oslo.utils==3.40.2
python-keystoneclient==3.18.0
python-swiftclient==3.6.0
rfc3986==1.2.0
stevedore==1.30.0
wrapt==1.11.1
